iteration

+++ SSU.DepManagement/SSU.DM.LogicLayer.Interfaces/Teachers/ITeacherLogic.cs
@@ -5,14 +5,14 @@ namespace SSU.DM.LogicLayer.Interfaces.Teachers;
 public interface ITeacherLogic
 {
     IReadOnlyList<TeacherViewItem> GetAll();
-    
+
     IReadOnlyList<TeacherCapacitiesViewItem> GetTeacherCapacities();
-    
-    void UpdateCapacities(long teacherId, Dictionary<long, int> totalHoursBySemester);
-    
+
+    void UpdateCapacities(long teacherId, Dictionary<long, float> totalHoursBySemester);
+
     void Create(TeacherViewItem viewItem);
-    
+
     void Update(TeacherViewItem viewItem);
-    
+
     void Delete(long id);
 }
